fix(focus): set focus based on DOM order
BREAKING CHANGE:
The `form-group` directive must be specified as a CSS class, rather than as an attribute. This allows the browser to find controls that have errors by
using document.querySelector('.form-control .ng-invalid').

Additional classes 'form-group-checkbox' (in the template) and 'form-group-radio' (must be added manually) can be added to the same element that has 'form-group' to support CSS styling.